requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
websockets==12.0
speechrecognition==3.10.0
pyttsx3==2.90
opencv-python==4.9.0.80
Pillow>=11.1.0
numpy>=1.26.4
python-multipart>=0.0.20